실수 단위 추가 및 버그 수정
1. supplierId로 Name을 가져오는 로직 수정
2. html에 step을 추가하고 실수형식으로 수정
3. 각 입력란의 기본값 생성

+++ stupping-mall/src/main/java/com/nerds/stuppingmall/dto/NotebookInfoResponseDto.java
@@ -23,10 +23,10 @@ public class NotebookInfoResponseDto {
 	private String cpuName;
 	private String gpuName;
 	private double weight;
-	private int screenSize;
+	private double screenSize;
 	private int ramSize;
 	private int ssdSize;
 	private int hddSize;
-	private int batterySize;
+	private double batterySize;
 	private Usage usage;
 }
